requirements.txt
aniso8601==9.0.1
black==23.3.0
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-RESTful==0.3.10
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
tomli==2.0.1
typing-extensions==4.6.3
Werkzeug==2.3.6
zipp==3.15.0